refactor: refactor login-page component path & style
