+++ src/app/page.tsx
@@ -1,5 +1,4 @@
-import LoginButton from '@/components/login-button'
-import SignIn from '@/components/sign-in'
+import LoginButton from '@/components/login-page/login-button'
 import { Button } from '@/components/ui/button'
 
 export default function Home() {
@@ -8,7 +7,6 @@ export default function Home() {
       <div>Pic me pls YOYO</div>
       <div>
         <Button>Button from shadcn</Button>
-        <SignIn />
       </div>
       <div className='flex flex-row space-x-2'>
         <div className='size-10 rounded-full bg-base-primary'></div>
