+++ src/components/login-page/index.tsx
@@ -2,13 +2,14 @@
 
 import { useState } from 'react'
 
+import { cn } from '@/lib/utils'
 import { UserType } from '@/type/user'
 import Image from 'next/image'
 
-import SignInDrawer from '@/components/sign-in-drawer'
-import SignInForm from '@/components/sign-in-form'
+import SignInDrawer from '@/components/login-page/sign-in-drawer'
+import SignInForm from '@/components/login-page/sign-in-form'
 
-import LoginButton from '../login-button'
+import LoginButton from './login-button'
 
 export default function LoginPageComponent() {
   const [userType, setUserType] = useState('')
@@ -29,76 +30,47 @@ export default function LoginPageComponent() {
 
   return (
     <div className='flex h-screen w-full'>
-      {!isSignIn ? (
-        <div className='flex w-full flex-col items-center justify-center'>
-          <div className='flex flex-col items-center justify-center space-y-4'>
-            <Image alt={''} src='logo.svg' width={64} height={67.75} />
-            <h1 className='text-base font-medium'>Sign in as</h1>
-            <p className='text-xs text-zinc-500'>
-              Sign in to book photographers or showcase your talent.
-            </p>
-            <div className='hidden flex-row items-center space-x-4 md:flex'>
-              <LoginButton
-                userType='Customer'
-                onClick={() => handleUserIconClick('Customer')}
-              />
-              <LoginButton
-                userType='Photographer'
-                onClick={() => handleUserIconClick('Photographer')}
-              />
-            </div>
-            <div className='flex flex-row items-center space-x-4 md:hidden'>
-              <SignInDrawer
-                userType='Customer'
-                onClick={() => handleUserIconClick('Customer')}
-                onClose={() => handleDrawerClose()}
-              />
-              <SignInDrawer
-                userType='Photographer'
-                onClick={() => handleUserIconClick('Photographer')}
-                onClose={() => handleDrawerClose()}
-              />
-            </div>
+      <div
+        className={cn(
+          'flex w-full flex-col items-center justify-center',
+          isSignIn && 'md:w-1/2'
+        )}
+      >
+        <div className='flex flex-col items-center justify-center space-y-4'>
+          <Image alt={''} src='logo.svg' width={64} height={67.75} />
+          <h1 className='text-base font-medium'>Sign in as</h1>
+          <p className='text-xs text-zinc-500'>
+            Sign in to book photographers or showcase your talent.
+          </p>
+          <div className='hidden flex-row items-center space-x-4 md:flex'>
+            <LoginButton
+              userType='Customer'
+              onClick={() => handleUserIconClick('Customer')}
+            />
+            <LoginButton
+              userType='Photographer'
+              onClick={() => handleUserIconClick('Photographer')}
+            />
+          </div>
+          <div className='flex flex-row items-center space-x-4 md:hidden'>
+            <SignInDrawer
+              userType='Customer'
+              onClick={() => handleUserIconClick('Customer')}
+              onClose={() => handleDrawerClose()}
+            />
+            <SignInDrawer
+              userType='Photographer'
+              onClick={() => handleUserIconClick('Photographer')}
+              onClose={() => handleDrawerClose()}
+            />
           </div>
         </div>
-      ) : (
-        <div className='flex w-full flex-col items-center justify-center md:w-1/2'>
-          <div className='flex flex-col items-center justify-center space-y-4'>
-            <Image alt={''} src='logo.svg' width={64} height={67.75} />
-            <h1 className='text-base font-medium'>Sign in as</h1>
-            <p className='text-xs text-zinc-500'>
-              Sign in to book photographers or showcase your talent.
-            </p>
-            <div className='hidden flex-row items-center space-x-4 md:flex'>
-              <LoginButton
-                userType='Customer'
-                onClick={() => handleUserIconClick('Customer')}
-              />
-              <LoginButton
-                userType='Photographer'
-                onClick={() => handleUserIconClick('Photographer')}
-              />
-            </div>
-            <div className='flex flex-row items-center space-x-4 md:hidden'>
-              <SignInDrawer
-                userType='Customer'
-                onClick={() => handleUserIconClick('Customer')}
-                onClose={() => handleDrawerClose()}
-              />
-              <SignInDrawer
-                userType='Photographer'
-                onClick={() => handleUserIconClick('Photographer')}
-                onClose={() => handleDrawerClose()}
-              />
-            </div>
-          </div>
-          {userType === 'Customer' ? (
-            <SignInForm userType='Customer' />
-          ) : userType === 'Photographer' ? (
-            <SignInForm userType='Photographer' />
-          ) : null}
-        </div>
-      )}
+        {userType === 'Customer' ? (
+          <SignInForm userType='Customer' />
+        ) : userType === 'Photographer' ? (
+          <SignInForm userType='Photographer' />
+        ) : null}
+      </div>
     </div>
   )
 }
