--- src/components/sign-in.tsx
@@ -1,16 +1,0 @@
-import { signIn } from '@/auth'
-
-import { Button } from '@/components/ui/button'
-
-export default function SignIn() {
-  return (
-    <form
-      action={async () => {
-        'use server'
-        await signIn('google')
-      }}
-    >
-      <Button type='submit'>Signin with Google</Button>
-    </form>
-  )
-}
